pyproject.toml
[tool.poetry]
name = "mis-project-api"
version = "2.2.0"
description = "Management Information System Project API"
authors = ["GigaWatte <wolfrage2580@gmail.com>", "m-riabchenko <mriabchenko11@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.2"
uvicorn = {extras = ["standard"], version = "^0.23.2"}
loguru = "^0.7.2"
python-multipart = "^0.0.6"
itsdangerous = "^2.1.2"
tortoise-orm = {extras = ["asyncpg"], version = "^0.21.7"}
yoyo-migrations = "^8.2.0"
apscheduler = "^3.10.4"
aio-pika = "^9.3.0"
ujson = "^5.8.0"
python-dotenv = "^1.0.0"
pydantic-settings = "^2.0.3"
python-jose = "^3.3.0"
redis = "^5.0.1"
dataclasses-json = "^0.6.1"
psycopg2-binary = "^2.9.9"
pymongo = {extras = ["srv"], version = "^4.6.0"}
motor = "^3.6.0"
bcrypt = "^4.1.2"
fastapi-pagination = "^0.12.21"
google-api-python-client = "^2.127.0"
aiohttp = "3.9.5"
aiohttp-socks = "^0.8.4"
fake-useragent = "^1.5.1"
aiodns = "^3.2.0"
gunicorn = "22.0.0"
cron-converter = "^1.2.1"


[tool.poetry.group.test.dependencies]
pytest = "^8.1.1"
httpx = "^0.27.0"
pytest-dotenv = "^0.5.2"
pytest-asyncio = "^0.23.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "-ra -q"
testpaths = [
    "tests"
]

filterwarnings = [
    "ignore::DeprecationWarning"
]
# log_cli = "true"
# log_cli_level = "DEBUG"

# log_file= ""./logs/tests/last_test.log"
# log_file_level = "DEBUG"

# asyncio_mode = auto

env_override_existing_values = 1
env_files = ".env"
